docs: arrange samples
