+++ package.json
@@ -138,6 +138,6 @@
     "node-fetch": "^2.6.0",
     "prettier": "^1.19.1",
     "react-test-renderer": "^16.13.0",
-    "react-zeroconfig": "^3.27.0-alpha.14"
+    "react-zeroconfig": "^3.27.0-alpha.15"
   }
 }
