+++ src/_packages/@handbook/components/components/Example.tsx
@@ -155,5 +155,10 @@ export const Example = styled(ExampleBase)`
 
   > :nth-child(2) {
     margin-top: 10px;
+
+    pre.prism-code {
+      overflow-y: auto;
+      max-height: 350px;
+    }
   }
 `;
