+++ src/_packages/@lunit/insight-viewer/__pages__/InsightViewer/Basic.mdx
@@ -4,7 +4,7 @@ import { InsightViewerPreview } from '@lunit/handbook';
 
 # `<InsightViewer>`
 
-<Example example={component('./Basic.sample')}>
+<Example example={component('./Basic/Sample')}>
   <InsightViewerPreview height={720}/>
 </Example>
