+++ src/_packages/@lunit/insight-viewer/__pages__/InsightViewer/Basic/Sample.tsx
